% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_report.R
\name{create_png_html_report}
\alias{create_png_html_report}
\title{创建基于PNG的HTML报告（修正转义问题）
创建基于PNG的HTML报告（英文版）}
\usage{
create_png_html_report(
  seg.tab,
  best_ploidy,
  best_cellularity,
  output_dir_name,
  img_dir,
  chr_images
)
}
\description{
创建基于PNG的HTML报告（修正转义问题）
创建基于PNG的HTML报告（英文版）
}
