% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_report.R
\name{generate_png_hrd_report}
\alias{generate_png_hrd_report}
\title{生成PNG图片的HRD分析HTML报告}
\usage{
generate_png_hrd_report(
  seg.tab,
  seqz_list,
  celluPloidyMat,
  best_ploidy,
  best_cellularity,
  output_dir_name = "./hrd_png_report"
)
}
\arguments{
\item{seg.tab}{the segment file.}

\item{seqz_list}{the raw input.}

\item{celluPloidyMat}{cellularity and ploidy matrix.}

\item{best_ploidy}{the optimized ploidy.}

\item{best_cellularity}{the optimized cellularity.}

\item{output_dir_name}{the output directory.}
}
\value{
HTML文件路径
}
\description{
这个版本先用你的原始函数生成PNG图片，然后在HTML中加载显示
}
