% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti.R
\name{opti_and_segs}
\alias{opti_and_segs}
\title{Optimized the model, get best cellularity and best ploidy, then segments each chromosome}
\usage{
opti_and_segs(
  data.file,
  number_of_cores = 6,
  CNt.max = 8,
  min.tumor.reads = 20,
  cellularity = seq(0.1, 1, by = 0.01 * grid_density),
  ploidy = seq(1, 6, by = 0.1 * grid_density)
)
}
\arguments{
\item{data.file}{the input file, can be got from sequenza-utils}

\item{number_of_cores}{the number of cores used to optimized the model.}

\item{CNt.max}{the upper limit of ploidy.}

\item{min.tumor.reads}{the mininal number of reads in tumor required to support the SNP.}

\item{cellularity}{the range of the input cellularity.}

\item{ploidy}{the range of the input ploidy.}
}
\value{
list of below elements.

\code{seg.tab} list of all segments.

\code{seqz_list} the input data.

\code{celluPloidyMa} cellularity and ploidy matrix.

\code{best_ploidy} optimized ploidy.

\code{best_cellularity} optimized cellularity.
}
\description{
Optimized the model, get best cellularity and best ploidy, then segments each chromosome
}
\examples{
data.file <-  system.file("extdata", "example.seqz", package = "HrdDetective");

seg_obj<-opti_and_segs(data.file=data.file);

output_plot(seg_obj$seg.tab, seg_obj$seqz_list, seg_obj$celluPloidyMat,
            seg_obj$best_ploidy,
            seg_obj$best_cellularity,
            output_dir_name="./");

}
