% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup_r.R
\name{pileup_from_bam}
\alias{pileup_from_bam}
\title{从BAM文件生成pileup统计}
\usage{
pileup_from_bam(
  bam_file,
  region = NULL,
  min_mapq = 0,
  min_base_quality = 0,
  min_depth = 0,
  min_af = 0.2,
  max_depth = 1e+06,
  distinguish_strands = FALSE,
  distinguish_nucleotides = TRUE,
  ignore_query_Ns = TRUE,
  include_deletions = FALSE,
  include_insertions = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{bam_file}{BAM文件路径}

\item{region}{基因组区域，格式为"chr:start-end"或GRanges对象}

\item{min_mapq}{最小MAPQ质量分数}

\item{min_base_quality}{最小碱基质量分数}

\item{min_depth}{最小深度}

\item{max_depth}{最大深度}

\item{distinguish_strands}{是否区分正负链}

\item{distinguish_nucleotides}{是否区分不同碱基}

\item{ignore_query_Ns}{是否忽略query序列中的N}

\item{include_deletions}{是否包含缺失}

\item{include_insertions}{是否包含插入}

\item{verbose}{是否显示详细信息}
}
\value{
包含pileup结果的DataFrame
}
\description{
从BAM文件生成pileup统计
}
