% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pileup_r.R
\name{pileup_whole_bam}
\alias{pileup_whole_bam}
\title{为整个BAM文件生成pileup}
\usage{
pileup_whole_bam(bam_file, verbose = TRUE, ...)
}
\arguments{
\item{bam_file}{BAM文件路径}

\item{...}{传递给pileup_from_bam的参数}
}
\value{
包含全基因组pileup的DataFrame
}
\description{
为整个BAM文件生成pileup
}
